resolve dependency cycle
